Made some bug fixes and added more validation
- Added more validation to the cart controller
- Add more validation to deleting product
- Add an 'is admin' attribute to the user database, model and seeder
- Deleted category controller since it wasn't necessary

+++ app/Http/Controllers/AdminProductController.php
@@ -95,7 +95,7 @@ class AdminProductController extends Controller
     
     public function destroy(Product $product)
     {
-        Storage::delete($product->product_image);
+        if($product->product_image != NULL) Storage::delete($product->product_image);
         Product::destroy($product->id);
         return redirect("/admin")->with("success", "Product has been deleted successfully");
     }
